Add rubocop and fix all existing violations
I've added `rubocop`, `rubocop-rails` and, as suggested the first time I
ran it, `rubocop-capybara`. I also enabled the NewCops as suggested.

With this suggested configuration, the auto-generated rails code had a
number of violations. I fixed these with `rubocop -A`. The remaining
violation was then `Style/Documentation`. I've decided to disable this
rule as I've never really added doc strings to a rails app and I don't
intend to start doing it here.