Live reload changes in development
I followed a blog post[1] to ensure that changes we make to the
application are automatically reloaded in the browser. Because we're
using tailwind we also need to add the `app/assets/builds` path to the
`hotwire_livereload` `listen_paths` since tailwind-rails is generating
the CSS on the fly based on the utility classes used in the views.

[1] https://dev.to/thomasvanholder/how-to-set-up-rails-hotwire-live-reload-38i9