Basic flow for registering interest in the project
I scaffolded an `interest` resource and then trimmed it down to just
create and new actions.

I've chosed to redirect to a "thank you" page following a successful
submission.

I had to switch to using the headless firefox driver for the generated
system tests because I was seeing a "Unable to find latest point
release version for 115.0.5790" locally when using the `:chrome`
driver. I might need to revist this if and when this starts to run in
a CI environment.