Don't save a new interest if we already have the email
It's possible that a user might decided to register interest more than
once with the same email address. I don't really want to display an
"you've already registered" notice since this would leak that
information for any email address pasted into the form. Instead, if we
find an existing registration, we'll proceed as if it were new.