Configure email sending in development/test
To be able to send emails in development we set a default host, and
use the letter_opener gem so they open in the browser as well as
display in the log output. For test we need to specify the host if we
want to generate urls.

It seems reasonable to exclude development.rb from the
Metrics/BlockLength cop - I'm not sure there's much to be gained yet
from splitting this up into smaller pieces to keep the cop happy.