Send a confirmation email when someone registers interest
The best way of validating that we have a real email address is to ask
the user to "opt-in" by sending them an email to confirm their
address. This also ensures we have explicit consent to contact
them[1].

We generate a confirmation token when a new Interest is complete, and
then adds a controller action that checks this token and confirms the
email address. We're somewhat relying on the obscurity of the token
here - someone would have to craft a valid token in order to confirm
another persons email address.

I've decided to keep the confirmation token after the Invitation has
been created, in case that person decides to sign up again.

For now, successful confirmation redirects to the generic "thank you"
path. I intend to work on some relevant copy for these pages later.

[1] https://ico.org.uk/for-organisations/direct-marketing-and-privacy-and-electronic-communications/guide-to-pecr/electronic-and-telephone-marketing/#consent