Configure development to send emails using Postmark
I set up an account with Postmark, and have created a `Music
Coop (development)` "server" (postmark's term for a folder or group of
settings).

The api key in the encrypted credentials.yml.enc is the API key for
this server.

At the moment, postmark will only send from and to the @gofreerange
domain. We'll need to configure the app to use its own domain, add a
new postmark server and get that server approved by postmark once we
want to start sending emails for real.