Use friendly ids for albums
I'd like to use friendly URLs for artists and albums, so that they are
more search engine friendly and shareable. For example:

    https://jam.coop/albums/the-white-album

Rather than

    https://jam.coop/albums/1

I've added the `friendly_id` gem[1] to achieve this. I removed the
CreateFriendlyIdSlugs migration that was added as part of the
installation for now - this allows slug histories to be versioned if,
for example, the album title changes. Since we don't have any
interface for changing the title yet, I haven't included this
additional complexity.

[1] https://github.com/norman/friendly_id