Generate a login/authentication system using authentication-zero
I ran

   rails generate authentication

then `bundle install` and `rails db:migrate`

I wanted to keep this commit as close as possible to the diff
generated by running this command, so that if, in the future the
`authentication-zero` gem is upgraded it might be possible to run it
again in the context of this commit to see what's changed, and how to
apply any changes. I had to make a couple of changes to the generated
code to support that:

1. There were some conflicts with the exisiting `test/test_helper.rb`
and `test/application_system_test_case.rb`. The purpose of these
changes seemed to be to add the `sign_in_as` method to both helpers,
so I just added that diff rather than let it clobber both files.
2. We already had a `root` set up in `routes.rb` so I commented out
the existing one so that the one added by this generator would work

Additionally I skipped the new authentication before action in our
InterestsController so that the system tests would continue to
pass. And I changed the `test/system/interests_test.rb` to visit the
register interest urls instead of the root url (which has now
changed).

The generated code is not valid according to our rubocop rules, so
I've added all of the generated code to the rubocop todo file and will
clean that up in a subsequent commit.