Add TranscodeJob to transcode track audio to mp3
This job takes a `track` and transcodes its original to MP3 V0 format.

To explain the shell out to `ffmpeg`

    ffmpeg
    -y                    # replaces output if it exists
    -nostats              # reduces output verbosity
    -loglevel 0           # reduces output verbosity
    -i #{input.path}      # sets the input file
    -codec:a libmp3lame   # uses the libmp3lame codec
    -q:a 0                # sets the quality to VBR[1]
    -f mp3                # explictly set output format if it can't be
                          # inferred from file name
    #{output.path}        # set output format

I've decided to make a real call to `ffmpeg` as part of the tests to
catch any errors introduced into the ffmpeg command switches. `ffmpeg`
is available in the Github Actions CI environment, and the input file
in the fixtures is very small so this runs fast.

[1] https://trac.ffmpeg.org/wiki/Encode/MP3