Don't try to add MP3 URL to tracklist items that don't have one
If a `track.preview` returns nil (for example if the transcode job
hasn't yet completed) we can't generate a URL for it. This commit
copes with that situation. It'll mean that these tracks are left out
of the "playlist" if a user hits play if they happen to visit before
the preview is generated.

At the moment this can only happen to the admin user - when we make
these pages public perhaps we can block "publishing" the album unless
it has previews generated for it.