A more presentable /artists
A grid of artist profile picture sorted alphabetically is a bit more
appealing than the previous text list.

The `w-full aspect-w-1 aspect-h-1` classes are a trick to make an
empty div fill a square. This happens if the artist does not have a
profile_picture and ensures that the grid layout doesn't look too
wonky.