Allow non signed-in users to visit artist and album pages
This commit allows non signed-in users to follow the current "buy an
album" journey starting at the artist index page, visiting an artist
page and finally visiting an album page.

The pages still have UI elements for adding and editing resources that
don't make sense to show to non-signed in users (since they will be
immediately redirected to sign in). I'll hide those in a subsequent commit.