Ensure Current.session is set on every request
Before this change we only set Current.session (that is, whether a
user is logged in and has a current session set) on calls to
`authenticate`. In order to allow us to differentiate between
signed-in and signed-out users in our controller actions and views we
need to perform this check on every request.

With this change the #authenticate method can just check for the
presence of a Current.user.