Remove redundant assert_selector in system test to avoid flakiness
I was seeing some flakiness with these selectors asserting the
presence of `h1` and `h2`. I suspect this might be due to the use of
Turbo to replace the page DOM rather than a full page reload. Since
the other assert_text calls confirm that we're on the correct page, I
think I can safely remove these without losing coverage.