Allow TranscodeJob to transcode to CBR MP3 (128kb/s)
The ffmpeg parameters for this file type are mostly the same as in
8f4110ee but with a constant bit rate of 128kb/s.

I think this file format will be useful as the default
streaming/previewing one as it results in a lower file size (meaning
faster delivery over slower connections). Since we don't have any
mechanism of stopping visitors downloading the files we use for
previewing it also provides an incentive for them to buy, since a
128kb/s MP3 will sound OK, but have some artifacts not present in our
better quality and larger transcodes.