Allow TranscodeJob to transcode to flac
The ffmpeg parameters for this file type are mostly the same as in
8f4110ee but using the flac codec.

Flac is a lossless audio codec so it converts the WAV file input into
a smaller file size without introducing any artifacts.

The content type for flac appears to be `audio/flac`[1].

Being able to download a lossless version of the audio is one of the
advantages of paying to download music rather than streaming it, so I
think it's worth supporting in this first iteration of downloads.

[1] https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Containers#flac