Transcode to all formats on track save
This ensures that the flac and mp3v0 formats are available to build
the download asset for the album.

This is naturally a bit wasteful of compute and storage resources and
we might want to reconsider this at some stage (e.g. by transcoding
on-demand when fans request a particular download format). But, for
now, this simplifies getting a download asset available to purchase so
we can reconsider if and when we have an issue.