Add ZipDownloadJob for creating downloadable albums
This job creates a zip file containing all of the transcoded files of
the given format.

I've taken a fairly brute force approach to this at the moment - we
first download all of the transcodes from the asset storage (i.e. S3
in production) to a temporary directory, then create a zip file from
these files using rubyzip. We then upload that zip file back to S3
attached to a new Download record. There's possibly a less memory
intensive way of doing this and I think there's a slight risk with
storing files in a temporary directory if machines are restarted
etc. But hopefully this will be robust enough to get us going.

We only need one download asset per transcode format, so we delete any
existing ones on creation of a new one.