Redirect to previous page on successful signin
When a user visits a page that requires authentication we want to
authenticate them and return them to their intended page.

I'm a bit concerned that we have an authentication-zero provided
Session as well as the rails `session` - this could be a bit
confusing. I'm not sure how to fix that yet.