Basic checkout flow
This commit adds a very basic checkout flow for purchasing an album.

The purchaser clicks "buy" on an album page, is shown a "confirmation"
page (which just contains a "checkout" button for now) and is then
redirected to stripe. On successful purchase they are redirected to
the album page, but no further indication is currently shown to
indicate their purchase.

The price of the album is currently hard-coded.

I've encapsulated the logic for creating a
Stripe::Checkout::Session (provided by the stripe-ruby gem) in a
StripeService class to make it easier to test and stub (to avoid
connecting to a third-party system when running tests).

In subsequent commits I'll flesh out this basic journey to store the
purchase and allow the purchaser to download the album.