Thank the customer and provide downloads on successful purchase
This commit completes the checkout flow by allowing the customer to
download the zip files of the album they have purchased.

The Purchase created has a uuid so is somewhat unguessable, but not
secure. We provide the URL for this purchase as the success_url when
creating the stripe checkout and assume that a successful redirect
from Stripe means we can allow the customer to download. We store the
album id when we create the purchase record so that we don't need to
fetch anything from Stripe in order to provide the download.