Allow a customer to name their price at checkout
Closes #24

This commit allows a fan to specify the price they wish to pay at
checkout, where the price must be equal to or greater than the
"suggested price" defined by Album#price.

The Stripe API requires the `unit_amount` to be an integer (pence in
this case) so I've added a conversion method `price_in_pence` to
Purchase.