Enable automatic tax calculation for purchases
We're registered to pay VAT in the UK so I've added a UK tax
registration in Stripe[1]. This config parameter tells stripe to
automatically calculate and include tax for anywhere we are tax
registered.

By default Stripe considers the price to be inclusive of VAT for
payments in GBP (so if the album is priced at £5 that will be what the
customer pays).

I think if we make sales within the EU we will have to turn on the
registrations for the countries we are registered in. For Mission
Patch I think we use some kind of "one stop shop" to pay taxes across
the EU, but we'll need to talk to Goddard to work out what to do once
we've made sales in any other countries.

[1] https://dashboard.stripe.com/tax/registrations