Suppress browserslist warning
The following warning was being displayed when running the tests and
when precompiling assets:

> Browserslist: caniuse-lite is outdated. Please run:
>  npx update-browserslist-db@latest
>  Why you should do it regularly: https://github.com/browserslist/update-db#readme

This warning is generated when the `tailwindcss` CLI command is run by
the "tailwindcss:build" rake task with which `tailwindcss-rails`
enhances the "test:prepare" & "assets:precompile" rake tasks.

Since this `tailwindcss` command is running code that is entirely
vendored and we don't even have a `package.json` file, following the
instructions to run `npx update-browserslist-db@latest` doesn't fix the
warning.

`BROWSERSLIST_IGNORE_OLD_DATA` is an environment variable in the
`browserslist` package which suppresses the warning and doesn't even
attempt to load a `package.json` file [1] seems to make sense. However,
this environment variable is _intentionally_ un-documented and the
suggestion is that it's a bad idea to use it [2]. However, since we have
no other way to fix the warning that seems a bit pointless and annoying.

I suspect the "right" way to fix this is to not use `tailwindcss-rails`
gem, but instead to install the `tailwindcss` npm package (and
associated packages) via a `package.json` file.  However, that feels
like a job for another day!

[1]: https://github.com/browserslist/browserslist/commit/0ae7155a6c7d8f58eaeb3ea450c21bec76de2763
[2]: https://github.com/browserslist/browserslist/pull/728