Indicate which tracks have been transcoded in album admin area
We're about to start adding a number of new albums to the site. Before
we hit "publish" on an album we want to know if all of the tracks have
been transcoded into the formats we support (so that previews and
downloads will work ok). I've also noticed a couple of out of memory
errors in production on the sidekiq instance, so having some debugging
information here about the transcodes will help warn us of that or
similar issues.