Add a listed boolean to Artist
We're about to add some new artists to the platform. Before anyone can
see them it'd be nice to share the link with the artist concerned to
make sure they're ok with all the data etc. To do that we'll add a
simple "listed" boolean which when "false" will exclude them from
/artists and other places they might surface. This doesn't provide
much security but it should be enough to stop them being accidentally
discovered in an early state - without the added complexity of
one-time URLs or user accounts etc.