Add track duration to track list
This commit adds the formatted time (in minutes:seconds) to the
tracklist. I think this helps a listener decide whether to listen and
also gives a better indication of what a download will contain if
purchased.

The ActiveStorage::AnalyzeJob which runs on each active storage upload
already has support for extracting the duration of an mp3 audio
file (in seconds) so we can use that to display the track time.

See: #39