Show album release date
We're currently adding albums to jam that have been released
elsewhere (on Bandcamp). We'd like to retain their "original" release
date, so that we can order them chronologically on the artist page.

This commit adds an optional `released_at` date field to album and
displays the release date between the "about" and "credits" section on
the album show page.

Things might get a little more complicated when we have albums that
are first published on jam and/or haven't been released elsewhere, but
we can cross that bridge when we come to it.

See: #42