Add a "add me to the mailing list" checkbox on purchase
At some point in the future we'd like artists to be able to contact
their fans directly. Although typically we'd avoid adding features to
the code until we need them, in this case staying compliant with GDPR
etc. means that it's easier if we gather that explicit opt-in at the
time of purchase rather than retrospectively have to seek it.

We don't collect the email address of the purchaser in our code (yet)
but it is collected as part of the stripe checkout form. When we do
need these emails we'll want to reconcile this boolean with the data
stored in stripe. To make that easier I'm going to add the stripe
session id to the purchase in a subsequent commit.