Indicate progression of uploads on albums/edit
I've wrapped the input field in a div to make it easier to find the
relevant progress indicator with `parentElement.querySelector`. I
considered adding a unique `id` to this `div` so that I could add them
to the controllers targets and then find the relevant ones in the
controller methods, but this seems to work ok so I'm going to live
with the slightly non-idiomatic style.