Allow new tracks to be added when editing/creating an album
This relies on `accepts_nested_attributes_for` automatically creating
new nested track records as long as a new id is passed in through the
forms parameters. For example, for the track title input field, the
`name` attribute is `album[tracks_attributes][:some_unused_id][title]`

We therefore create a `template` (the HTML template tag is hidden by
default) with the appropriate `fields_for` a new track, passing in the
explicit string `__CHILD_INDEX__`. The javascript in the stimulus
controller triggered when the "add track" link is clicked replaces
this string with a integer derived from the current time.

I got inspiration for this approach from a detailed stack overflow
answer[1]. As we're already using stimulus controllers for this kind
of thing, I prefer this approach to some of the others suggested in
there.

[1] https://stackoverflow.com/questions/71713303/rails-7-dynamic-nested-forms-with-hotwire-turbo-frames