Do not send emails to suppressed addresses
I decided that the `EmailSubscriptionChange` model was overkill and was
going to be awkward to use with the `Interest` model as well as the
`User` model, so I've deleted the model and dropped the table and
replaced it with a single `sending_suppressed_at` column on both the
`user` table and the `interests` table. There are only two records in
the production database and so I decided it wasn't worth trying to
migrate the data.

The `EmailSubscriptionChangesController` now looks for both a `User` and
an `Interest` with the relevant email address and it will update the
`sending_suppressed_at` attribute accordingly.

I've also added guard conditions to all the existing mailer methods to
ensure we don't email any addresses that have been suppressed. I think
there's probably a nicer way of avoiding the duplciation (perhaps using
a before_action callback in the ApplicationMailer), but the way we're
using `params` in some mailers and method arguments in others rather
threw me!

We'll need to set up a separate webhook for the "transactional" Postmark
stream. This will be pretty much identical to the existing one for the
"broadcast" stream.

We also still need to update the production database based on the email
addresses that are already suppressed in our Postmark account.

Sorry it's rather a big commit!