Write the first newsletter
I've used a very basic HTML template based on a cheat-sheet I
found[1]. I think reliable cross-client HTML email is a bit of a
minefield and we might want to look for a better template at some
point. I've sent myself a copy and confirmed that it displays ok in
GMail.

We should be able to send this with:

    Interest.where(email_confirmed: true).each do |user|
      BroadcastMailer.newsletter(user).deliver_now
    end

[1] https://devhints.io/html-email