Basic flow for inviting a new user
We want to allow artists to administer their own releases/pages
etc. At least initially we don't want this to be a free-for-all, we'll
invite users to admin artists that we have created in the system.

As a first step this commit allows signed in users to invite new
users. The mechanism for this (creating them a user account with a
password and then sending a password reset) is cribbed from the
authentication-zero gem's `--invitable` generator, in particular the
InvitationsController[1].

I've set an expiry time of 1 day on this password reset token, as it
will probably take a user a bit longer to respond than if they
requested the password reset themselves.

I think we can definitely make some improvements to the usability of
this flow, but for now I think this is sufficient.

An obvious problem with this as it stands is that all users created
have the same permissions, so we'll need to add some kind of
role-based system to fix that.

[1] https://github.com/lazaronixon/authentication-zero/blob/0eb2ba99c4b46b856066bcbc1d1ece42bd87887b/lib/generators/authentication/templates/controllers/html/invitations_controller.rb.tt