Associate artists with users
We want an invited user to be able to administer one or more
artists. This migration sets up the necessary referrences. It's not
clear to me that we would want to destroy an artists and all their
releases if a user model was destroyed, so I'm kicking that can down
the road a bit with `restrict_with_exception`.