Introduce Pundit for authorisation
We want to allow admin users and regular/artist users to do different
things. I've decided to use Pundit to encapsulate these authorisation
concerns as I'm familiar with it from another project.

I ran the `pundit:install` generator and added

   after_action :verify_authorized

to the top of the ApplicationController. This causes each controller
action to raise an exception if the `authorize` method hasn't been
called. I then implemented authorization rules for each controller in
turn to maintain the existing application behaviour:

  - the two existing users (James and Chris) in the database are
    `admin` users set and can perform all of the existing actions
  - new users invited using the invitation mechanism in previous
    commits will not have the admin boolean. They, for now, can only
    view the session index, home page and edit/update their passwords.

To keep this commit somewhat manageable it does not hide any content
in the views from non-admin users. We'll do that in a subsequent
commit. We can also tighten up the `index` logic using Pundit's
scopes.