Add an navbar with avatar menu for logged in users
We need a convenient place to store action links for users. Currently
we have the /home page, but this isn't very easy to find. I think this
menu provides a nice place to store these links and adds a little
additional visual interest to the top of the page.

The implementation of this toggle-able menu closely follows Adam
Wathan's tutorial[1]. In particular the full-screen hidden
button (bound to the `avatar-menu#close` action) is used to dismiss
the menu by clicking outside of its bounds. I've also added a
`avatar-menu#escape` function bound to the `keydown` event on the
global `document` scope (with `keydown@document`) to provide an
accessible way of dismissing it.

I've pulled in a users "gravatar" if available[2] passing the `?d=mp`
option[3] to have a fallback "mystery person" if the user doesn't have
a gravatar set. We may want to improve this with a profile image
upload later.

[1] https://www.youtube.com/watch?v=If6XdbBQqmY
[2] https://docs.gravatar.com/general/hash/
[3] https://docs.gravatar.com/general/images/