Remove Home page
The links on this page are now contained in the avatar menu, so we no
longer need this page. Where we were previously redirecting to this
page, we're now redirecting to the `root_path` instead. I've added a
flash notice to that page to capture any messages that were previously
shown on home.

Note with this commit we lose a direct link to Sessions#index. I think
it's helpful to have this page around for debugging purposes, but I
don't know if we need to link "regular" users to it for now.