Fix bug preventing artist users from viewing other albums
Before this commit a logged in user who had associated artists could
only view their albums on their own artist show page due to the
`user.artists.any?` condition in the AlbumPolicy scope. I don't think
a Pundit scope is the correct way of modelling this as it's not
recommended to pass in additional context (here we need to know the
artist that the album scope is associated with). Instead I've
introduced a new `view_unpublished_albums?` scope in ArtistPolicy and
used that in the view to add additional albums.