Use publication state of albums to decide whether artist is listed
Previously we were using the listed boolean on Artist so that an admin
could create an artist with some published albums and then share the
URL of the artist page with the artist to make sure they were ok with
it. At that time we didn't want the artist to appear on /artists until
they gave us the OK.

Now artists can manage their own releases and access their artist page
from their avatar menu we no longer need rely on the "listed" boolean
on Artist to decide whether to include the artist on /artists.

I've decided to still show unlisted artists to admins on /artists and
indicate them with the `(unlisted)` text so that we can more easily
find them to perform admin actions.

This commit makes the `listed` column on `artists` redundant so I'll
remove it in the next commit.