Set Purchase#customer_email, #completed on successful checkout
This controller listens to the stripe webhook's
checkout.session.completed event and updates the customer_email and
completed fields on the corresponding Purchase record. This will allow
us to send an email to the customer and the artist on purchase, as
well as eventually associate purchases with their corresponding User
account.

I've followed the Stripe docs[1] in the structure of this webhook
code. It hasn't been secured using a token yet. The fixture file is
taken from a transaction I completed in development using the Stripe
CLI to forward webhooks to the locally running rails process.

[1] https://stripe.com/docs/webhooks