Secure Stripe webhook with endpoint_secret
I created a webhook in the production stripe account[1] and encrypted
the generated secret in the production.yml.enc.

I've decided to test this controller by expectation, using the
previous call to `Stripe::Event.construct_from` in the stubbed
response to the new `Stripe::Webhook.construct_event`. The value of
ENDPOINT_SECRET in test is `nil`.

In development when forwarding stripe events to localhost using the
Stripe CLI the command

    stripe listen --forward-to localhost:3000/stripe_webhook_events

outputs a secret that can be used locally for that session. For now I
think commenting out the ENDPOINT_SECRET line and setting it with the
generated one is probably enough to test things end to end, but it
might be good to make this a configurable value if that becomes
annoying.

[1] https://dashboard.stripe.com/webhooks/we_1ODPPOLj0fa0S8sdMnjJSMfV