Move processing of checkout.session.completed event to job
Stripe's Webhook best practices docs[1] suggest that the webhook
endpoint should return its status as quickly as possible. Since there
are things that could take time or go wrong on our side even though
we've received a correct hook, I've moved the logic for processing the
hook into a background job so that it doesn't delay returning a status
code to Stripe.

[1] https://stripe.com/docs/webhooks#best-practices