Email the customer when the purchase is completed
Customers currently get redirected to the purchase page when they have
completed the stripe checkout process. But if they navigate away from
this page they may struggle to find their downloads again. This commit
emails them a link to that page on successful purchase completion.

We can't guarentee that stripe webhooks are only received once so this
may result in the email being sent more than once. I'm not going to
address that in this commit.

Closes #63