Introduce Album#publication_status
We want to allow artists to request that their album is published. To
do this I've decided to introduce a publication_status enum with 0 =
draft and 1 = published. The migration sets the value of this enum
based on the current `published` boolean. I'll replace the current use
of the `published` boolean in subsequent commits.