Allow artist to put their own album into pending publication state
At the moment artists have to email us when their albums are ready to
publish on the site. This allows them to request publication with a
button click.

At the moment there's no way of letting the admins know when that has
happened, but I'll add that in a subsequent commit.