Specify node version (v20.9.0)
We're using a recent version of Ruby (v3.2.2), so it makes sense to me
that we use a recent version of JavaScript like this.

It looks as if Render was defaulting to use v14.17.0 [1], so this is
quite a jump, but as far as I can see it should support tailwindcss v3.
What's more the tests all pass locally and the app seems to run OK
locally. Note that Render is supposed to respect the `.node-version`
file.

Although the README mentions a `.tool-versions` file, it doesn't appear
to be in the repo. And in any case, I think using `.ruby-version` &
`.node-version` is more "portable" and `asdf` can be configured to
respect them using its `legacy-version-file` option [2].

[1]: https://render.com/docs/node-version
[2]: https://asdf-vm.com/manage/configuration.html#legacy-version-file