Update ArtistPolicy to allow users to create artists
For now I've decided to limit the number of artists a single user can
create to 2. I think we still want a bit of control over what's
happening and time to see how the feature is being used. There is one
user already who has 2 associated artists (that we've added manually)
so this number has precedence.