Reduce likelihood of artist slug clashes
Now we allow users to create their own artists there's an increased
chance that a user will select an artist name that is already in
use. Rather than immediately fallback to the `name-uuid` slug format
we first try the `name-music` and `name-sounds` candidates.

There's a chance that a user won't be happy with the slug with picked
for them and want to change it, but we can think about how to best
support that when it happens. For now an admin could go in to the
console and make things better.