Make S3 bucket publicly accessible
This is necessary if we want to use it as the origin for a Cloudfront
distribution.

Previously I've only needed to set `publicReadAccess` to `true`, but
that gave me an error and I found I needed to also set
`blockPublicPolicy`, although I may have set more of the sub-options to
`false` than I needed to.