Add custom domain & SSL cert for CF distribution
I've hard-coded the production domain as `music.gfr.coop` and the
development domain as `music-dev.gfr.coop`.

SSL certificates can only be created in the `us-east-1` region. This in
turn means we need to make use of the experimental
`crossRegionReferences` feature. However, this is almost exactly the
scenario used as the example in the documentation example [1].

This relies on manually adding a DNS CNAME record in the middle of the
CDK deployment in order to verify the SSL certificate.

Note that we could probably completely automate this if we used AWS
Route 53 to manage the DNS records for the domain.

[1]: https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib-readme.html#accessing-resources-in-a-different-stack-and-region