Add SSL certificate for CloudFront distribution
This uses DNS validation and so when deploying one of the MusicCoop
stacks, in order to see the details for the DNS CNAME record, you'll
either need to use the verbose option (e.g. `cdk deploy -v
MusicCoopDevelopmentStack`); or use the AWS web console to view the
certificate in the AWS Certificate Manager.

I've chosen to use the DNS validation method, becuase it means the SSL
certificates will be automatically renewed as long as the relevant DNS
CNAME record remains in place.