Render terms and condition markdown using appropriate style
Given we use tailwind I think it's easier to maintain this page (and
similar) if we author it in markdown and then render it with the
approprite utility classes in the markup.

To do that I've followed the approach I found in a blog post[1] and
defined a template handler which recognises `md` files and then
renders them with the new `TailwindMarkdownRenderer`. I've inherited
from Redcarpets default HTML renderer and redefined classes we use in
the terms and conditions doc (`h1-4` tags, `p`, `ul` and `pre`) to add
styles.

I've added this renderer file to the tailwind config so that any
tailwind styles we use in there are picked up and added to the bundled
styles.

I suspect there might be some performance implications to rendering
some of this markdown using ruby rather than Redcarpets optimised
renderer but I hope we'll be able to cache our way around that at a
later date if we need to.

[1] https://www.writesoftwarewell.com/how-to-render-markdown-views-in-rails/