Rewrite Terms to be more applicable to jam
- Replace Ampled with jam.coop in Terms
- Clarify that jam.coop is provided by Go Free Range and fix up our
business address.
- Clarify that the Peer Production license applies to this page. We retain this license
as it was attached to ampled's original.
- I've removed the bit about quotes and cartoons as we haven't retained
those from ampled's original.
- Remove reference to cookie, privacy and community policies as we don't
have those.
- Remove section 4a. This doesn't apply to us as we sell individual downloads rather than
offer subscriptions.
- Remove references to supporters/support as we have customers rather
  than the subscribers ampled has
- Remove reference to account settings for deleting artists. We don't
  support that yet.
- Prefer "Users" to "Members" to avoid confusion (GFR has members).
- Remove reference to direct damages. I'm not sure this makes sense
for us since we're not in the US. It might be worth finding an example
of T&Cs that are UK based to see if they set a liability limit like
this.
- Remove Dispute Resolution and Governing Law. This referred quite
heavily to New York law, which certainly doesn't apply to us. We may
need something comparable at some stage, but I think it's implied that
we'd resolve these things under UK law given our location.

Closes: #86