Switch ActiveStorage config to proxy mode
This means that all files stored using ActiveStorage will be served via
the Rails app, rather than (in redirect mode) where files are served by
redirecting to an S3 URL [1]. Obviously this will mean more traffic has
to pass through the Rails app server, but we need it in order to support
accessing the files via a CDN which will limit that traffic.

Note that if we decide we want to continue to serve _some_ ActiveStorage
files via redirect mode, then we'll need to remove this configuration
and use the `rails_storage_redirect_url` and/or
`rails_storage_proxy_url` route helpers on a case-by-case basis.

[1]: https://guides.rubyonrails.org/v7.1.1/active_storage_overview.html#proxy-mode