Change URLs for ActiveStorage files to use CDN
The new code in `config/routes.rb` is very closely based on the example
code in the "5.2.1 Putting a CDN in Front of Active Storage" section [1]
of the ActiveStorage guide. The only changes I've made are to change the
name of the new route helper method, to use Rails config rather than an
environment variable for the CDN host, and the formatting of the code.

The new route helper (`#cdn_url`) is then used to generate CDN versions
of URLs for each ActiveStorage file that appears in the app. Hopefully I
haven't missed any! If I have, it wouldn't be the end of the world,
because they will still work, they just won't be cached by the CDN, but
proxied via the Rails app which might be a bit slow.

The way the domains are set up is as follows:

In production:

* Non ActiveStorage requests: jam.coop -> Rails app
* ActiveStorage requests: cdn.jam.coop -> jam.coop -> Rails app -> S3 bucket

In development:

* Non ActiveStorage requests: dev.jam.coop -> Rails app
* ActiveStorage requests: cdn-dev.jam.coop -> dev.jam.coop -> Rails app -> S3 bucket

Note that you'll need to use something like ngrok in development to
expose your Rails app at dev.jam.coop. You'll also need to make the
following changes in `config/environments/development.rb`:

    config.active_storage.service = :amazon
    config.hosts << 'dev.jam.coop'

[1]: https://guides.rubyonrails.org/v7.1.1/active_storage_overview.html#putting-a-cdn-in-front-of-active-storage