Add maintenance mode using rack-maintenance
The `rack-maintenance` [1] middleware is configured to display a basic
maintenance page if the `MAINTENANCE` environment variable is set, i.e.
it doesn't rely on a file being written (filesystem on render.app is
read-only by default) or on a value in the database (which would
somewhat defeat the object of a maintenance mode.

I've mostly stolen HTML & CSS for `public/maintenance.html` from the
`turnout` gem [2].

[1]: https://github.com/tilsammans/rack-maintenance
[2]: https://github.com/biola/turnout/blob/48d51211a7c32a2050c726ef6c50d55052aa3574/public/maintenance.html